+++ test/io_tests/request_response_tests/session_test.py
@@ -34,10 +34,10 @@ class TestSession:
         with patch("logging.Logger.warning") as mock_warning:
             session.response_rec_callback(wrong_knxipframe, HPAI(), None)
             mock_warning.assert_called_with("Could not understand knxipframe")
-            assert session.success is False
+            assert session.response is None
 
         # Correct Response KNX/IP-Frame:
         res_knxipframe = KNXIPFrame.init_from_body(SessionResponse(secure_session_id=5))
         session.response_rec_callback(res_knxipframe, HPAI(), None)
-        assert session.success is True
+        assert session.response is not None
         assert session.response.secure_session_id == 5